fix: move support form to /contact and redirect /support
