+++ apps/site/next.config.ts
@@ -175,6 +175,16 @@ const nextConfig: NextConfig = {
         destination: '/:locale/research',
         permanent: true,
       },
+      {
+        source: '/:locale(en|es)/support',
+        destination: '/:locale/contact',
+        permanent: true,
+      },
+      {
+        source: '/support',
+        destination: '/contact',
+        permanent: true,
+      },
       // `/careers/index` is a real route (`careers/index/page.tsx`); do not add a redirect to `/careers`.
       // Handles requests to go.toddagriscience.com/invite
       {
