+++ apps/site/src/app/(unauthenticated)/[locale]/(contact)/contact/page.tsx
@@ -4,634 +4,837 @@
 
 import { FadeIn } from '@/components/common';
 import FormErrorMessage from '@/components/common/form-error-message/form-error-message';
-import { LegalSubtext } from '@/components/common/legal-subtext/legal-subtext';
 import MarketingGradientBox from '@/components/common/marketing-gradient-box/marketing-gradient-box';
 import SubmitButton from '@/components/common/utils/submit-button/submit-button';
 import { Button } from '@/components/ui';
-import {
-  Carousel,
-  CarouselApi,
-  CarouselContent,
-  CarouselItem,
-} from '@/components/ui/carousel';
-import {
-  Field,
-  FieldLabel,
-  FieldLegend,
-  FieldSet,
-} from '@/components/ui/field';
+import { Field, FieldGroup, FieldLabel, FieldSet } from '@/components/ui/field';
 import { Input } from '@/components/ui/input';
-import { submitContactToSheets } from '@/lib/actions/googleSheets';
-import isWorkEmail from '@/lib/utils/is-work-email';
+import { Textarea } from '@/components/ui/textarea';
+import { formatActionResponseErrors } from '@/lib/utils/actions';
 import { ErrorMessage } from '@hookform/error-message';
 import { zodResolver } from '@hookform/resolvers/zod';
-import Fade from 'embla-carousel-fade';
 import { useTranslations } from 'next-intl';
-import Link from 'next/link';
-import { useRouter } from 'next/navigation';
-import { useEffect, useState } from 'react';
+import { useRouter, useSearchParams } from 'next/navigation';
 import { useForm } from 'react-hook-form';
-import { ContactFormData, contactFormSchema } from './types';
+import { submitPublicInquiry } from './actions';
+import { PublicInquiryData, publicInquirySchema } from './types';
 
+/**
+ * Contact page. Allows users to submit a public inquiry support ticket.
+ *
+ * @returns The contact page with related logic.
+ */
 export default function Contact() {
-  const t = useTranslations('contactPage');
-  const [slide, setSlide] = useState(0);
   const router = useRouter();
-  const [totalSlides, setTotalSlides] = useState(-1);
-  const [api, setApi] = useState<CarouselApi>();
   const {
     register,
-    trigger,
-    getValues,
-    clearErrors,
-    setValue,
-    watch,
-    formState: { errors, isValid },
-  } = useForm<ContactFormData>({
+    handleSubmit,
+    reset,
+    setError,
+    formState: { errors, isSubmitting, isSubmitted, isValid },
+  } = useForm<PublicInquiryData>({
     defaultValues: {
-      name: undefined,
-      firstName: '',
-      lastName: '',
-      farmName: '',
-      email: '',
-      phone: '',
-      website: undefined,
-      isOrganic: undefined,
-      isHydroponic: undefined,
-      producesSprouts: undefined,
-      instagramHandle: '',
+      name: '',
+      lastKnownEmail: '',
+      response: '',
     },
-    resolver: zodResolver(contactFormSchema),
+    resolver: zodResolver(publicInquirySchema),
+    mode: 'onChange',
   });
+  const t = useTranslations('contactPage');
+  const searchParams = useSearchParams();
+  const intent = searchParams.get('intent');
+  const emailType = intent === 'forgot-email' ? '.forgotEmail' : '.email';
 
-  const {
-    firstName,
-    lastName,
-    farmName,
-    email,
-    phone,
-    website,
-    isOrganic,
-    isHydroponic,
-    producesSprouts,
-    instagramHandle,
-  } = watch();
-  const requiresWebsite = !isWorkEmail(email ?? '');
-  const organicSlideIndex = requiresWebsite ? 2 : 1;
-  const hydroponicSlideIndex = organicSlideIndex + 1;
-  const sproutsSlideIndex = organicSlideIndex + 2;
-  const instagramSlideIndex = organicSlideIndex + 3;
-  const canAdvance =
-    (slide === 0 &&
-      Boolean(firstName && lastName && farmName && email && phone)) ||
-    (requiresWebsite && slide === 1 && Boolean(website)) ||
-    (slide === organicSlideIndex && isOrganic !== undefined) ||
-    (slide === hydroponicSlideIndex && isHydroponic !== undefined) ||
-    (slide === sproutsSlideIndex && producesSprouts !== undefined) ||
-    (slide === instagramSlideIndex && Boolean(instagramHandle));
+  const onSubmit = async (data: PublicInquiryData) => {
+    const formData = new FormData();
+    formData.set('name', data.name);
+    formData.set('lastKnownEmail', data.lastKnownEmail);
+    formData.set('response', data.response);
 
-  useEffect(() => {
-    if (!api) return;
-
-    setTotalSlides(api.slideNodes().length);
-    setSlide(api.selectedScrollSnap());
-
-    api.on('select', () => setSlide(api.selectedScrollSnap()));
-    api.on('slidesChanged', () => setTotalSlides(api.slideNodes().length));
-  }, [api]);
-
-  function handleBack() {
-    api?.scrollPrev();
-  }
-
-  async function handleNext() {
-    if (!api) return;
-
-    const fieldsToValidate =
-      slide === 0
-        ? ([
-            'firstName',
-            'lastName',
-            'farmName',
-            'email',
-            'phone',
-          ] as (keyof ContactFormData)[])
-        : requiresWebsite && slide === 1
-          ? (['website'] as (keyof ContactFormData)[])
-          : slide === organicSlideIndex
-            ? (['isOrganic'] as (keyof ContactFormData)[])
-            : slide === hydroponicSlideIndex
-              ? (['isHydroponic'] as (keyof ContactFormData)[])
-              : slide === sproutsSlideIndex
-                ? (['producesSprouts'] as (keyof ContactFormData)[])
-                : slide === instagramSlideIndex
-                  ? (['instagramHandle'] as (keyof ContactFormData)[])
-                  : [];
-
-    const isStepValid = fieldsToValidate.length
-      ? await trigger(fieldsToValidate)
-      : true;
-
-    if (isStepValid) {
-      api.scrollNext();
-      clearErrors();
-    }
-  }
-
-  // This function is triggered directly by the <SubmitButton/> because onSubmit via the form itself isn't working for some reason.
-  async function onSubmit() {
-    await trigger();
-
-    if (isValid) {
-      const params = new URLSearchParams({
-        first_name: getValues().firstName,
-        last_name: getValues().lastName,
-        farm_name: getValues().farmName,
-        email: getValues().email,
-        phone: getValues().phone,
-        instagram_handle: getValues().instagramHandle || '',
+    try {
+      await submitPublicInquiry(formData);
+      reset();
+    } catch (error) {
+      const formatted = formatActionResponseErrors(error);
+      setError('root', {
+        message: formatted[0] ?? 'Something went wrong. Please try again.',
       });
-      router.push(`/signup?${params.toString()}`);
     }
-  }
-
-  async function onInstagramSubmit() {
-    await trigger();
-
-    if (isValid) {
-      const params = new URLSearchParams({
-        first_name: getValues().firstName,
-        last_name: getValues().lastName,
-        farm_name: getValues().farmName,
-        email: getValues().email,
-        phone: getValues().phone,
-        instagram_handle: getValues().instagramHandle || '',
-      });
-      // Push to Google Sheets
-      try {
-        await submitContactToSheets(Object.fromEntries(params.entries()));
-      } catch (error) {
-        console.error('Failed to submit to Google Sheets:', error);
-      }
-      // Redirect to What We Do
-      router.push(`/research`);
-    }
-  }
-
-  const isMatch =
-    (isWorkEmail(getValues().email) || getValues().website) &&
-    (getValues().isOrganic || getValues().isOrganic === undefined) &&
-    !getValues().isHydroponic &&
-    !getValues().producesSprouts;
+  };
 
   return (
     <main>
       <div className="max-w-[1400px] mx-auto px-15 lg:px-16 flex items-center justify-center">
-        <div className="grid grid-cols-1 md:grid-cols-2 gap-8 lg:gap-16 place-items-start mx-auto mt-5 md:mt-15 w-full max-w-[1200px] mx-auto">
+        <div className="grid grid-cols-1 md:grid-cols-2 gap-8 lg:gap-16 place-items-start mx-auto mt-25 md:mt-15 w-full max-w-[1200px] mx-auto">
           <MarketingGradientBox />
           <div className="flex w-full max-w-[530px] lg:max-w-none flex-col md:mr-0 lg:mr-10">
-            <form
-              onSubmit={(e) => e.preventDefault()}
-              onKeyDown={(e) => {
-                if (e.key === 'Enter' && isValid) {
-                  e.preventDefault();
-                  api?.scrollNext();
-                }
-              }}
-            >
-              <Carousel
-                setApi={setApi}
-                className="mx-auto flex flex-col justify-start w-full max-w-[300px] sm:max-w-[450px] md:max-w-[500px]"
-                plugins={[Fade()]}
-                opts={{ duration: 30, watchDrag: false }}
-              >
-                <CarouselContent className="p-0 h-auto">
-                  {/* Get Started with Todd*/}
-                  <CarouselItem>
-                    <FieldSet className="flex flex-col gap-3">
-                      <FieldLegend>
-                        <h1 className="text-2xl mb-8 md:mb-2 mt-10 text-left">
-                          {t('description')}
-                        </h1>
-                      </FieldLegend>
-                      {/** Honeypot */}
-                      <Field className="hidden">
-                        <FieldLabel htmlFor="name" className="leading-tight">
-                          {t('fields.name')}
-                        </FieldLabel>
-                        <Input
-                          className="border-[#848484]/80 border-1 "
-                          placeholder={t('placeholders.name')}
-                          type="text"
-                          {...register('name')}
-                        />
-                      </Field>
-
-                      <div className="grid grid-cols-1 sm:grid-cols-2 gap-2 sm:gap-4">
-                        <Field>
-                          <div className="flex flex-row justify-between">
-                            <FieldLabel
-                              htmlFor="first-name"
-                              className="leading-tight mb-[-6px]"
-                            >
-                              {t('fields.firstName')}
-                            </FieldLabel>
-                            <ErrorMessage
-                              errors={errors}
-                              name="firstName"
-                              render={({ message }) => (
-                                <FormErrorMessage errorMessage={message} />
-                              )}
-                            />
-                          </div>
-                          <Input
-                            className="border-[#848484]/80 border-1"
-                            required
-                            {...register('firstName', {
-                              required: 'This field is required.',
-                            })}
-                          />
-                        </Field>
-
-                        <Field>
-                          <div className="flex flex-row justify-between">
-                            <FieldLabel
-                              htmlFor="last-name"
-                              className="leading-tight mb-[-6px]"
-                            >
-                              {t('fields.lastName')}
-                            </FieldLabel>
-                            <ErrorMessage
-                              errors={errors}
-                              name="lastName"
-                              render={({ message }) => (
-                                <FormErrorMessage errorMessage={message} />
-                              )}
-                            />
-                          </div>
-                          <Input
-                            className="border-[#848484]/80 border-1"
-                            required
-                            {...register('lastName', {
-                              required: 'This field is required.',
-                            })}
-                          />
-                        </Field>
-                      </div>
-
-                      <Field>
-                        <div className="flex flex-row justify-between">
-                          <FieldLabel
-                            htmlFor="farm-name"
-                            className="leading-tight mb-[-6px]"
-                          >
-                            {t('fields.farmName')}
-                          </FieldLabel>
-                          <ErrorMessage
-                            errors={errors}
-                            name="farmName"
-                            render={({ message }) => (
-                              <FormErrorMessage errorMessage={message} />
-                            )}
-                          />
-                        </div>
-                        <Input
-                          className="border-[#848484]/80 border-1"
-                          required
-                          {...register('farmName', {
-                            required: 'This field is required.',
-                          })}
-                        />
-                      </Field>
-
-                      <Field>
-                        <div className="flex flex-row justify-between">
-                          <FieldLabel
-                            htmlFor="email"
-                            className="leading-tight mb-[-6px]"
-                          >
-                            {t('fields.email')}
-                          </FieldLabel>
-                          <ErrorMessage
-                            errors={errors}
-                            name="email"
-                            render={({ message }) => (
-                              <FormErrorMessage errorMessage={message} />
-                            )}
-                          />
-                        </div>
-                        <Input
-                          className="border-[#848484]/80 border-1"
-                          type="email"
-                          required
-                          {...register('email', {
-                            required: 'This field is required.',
-                          })}
-                        />
-                      </Field>
-
-                      <Field>
-                        <div className="flex flex-row justify-between">
-                          <FieldLabel
-                            htmlFor="phone"
-                            className="leading-tight mb-[-6px]"
-                          >
-                            {t('fields.phone')}
-                          </FieldLabel>
-                          <ErrorMessage
-                            errors={errors}
-                            name="phone"
-                            render={({ message }) => (
-                              <FormErrorMessage errorMessage={message} />
-                            )}
-                          />
-                        </div>
-                        <Input
-                          className="border-[#848484]/80 border-1"
-                          type="tel"
-                          required
-                          defaultValue={'+1'}
-                          {...register('phone', {
-                            required: 'This field is required.',
-                          })}
-                          maxLength={10}
-                        />
-                      </Field>
-                    </FieldSet>
-                    <div className="mt-10">
-                      <p className="text-xs mb-5 font-thin">
-                        Already started?{' '}
-                        <Link href="/login" className="underline font-normal">
-                          Login to complete your application
-                        </Link>
-                      </p>
-                    </div>
-                  </CarouselItem>
-                  {/* Website Question */}
-                  {!isWorkEmail(getValues().email) && (
-                    <CarouselItem>
-                      <FieldSet>
-                        <Field>
-                          <div className="flex flex-row justify-between mt-23 md:mt-16 lg:mt-12">
-                            <FieldLabel
-                              htmlFor="website"
-                              className="text-lg leading-tight"
-                            >
-                              {t('questions.website', {
-                                farm: getValues().farmName,
-                              })}
-                            </FieldLabel>
-                            <ErrorMessage
-                              errors={errors}
-                              name="website"
-                              render={({ message }) => (
-                                <FormErrorMessage errorMessage={message} />
-                              )}
-                            />
-                          </div>
-                          <Input
-                            placeholder="https://"
-                            className="border-[#848484]/80 border-1"
-                            {...register('website')}
-                            required
-                            {...register('website', {
-                              required: 'This field is required.',
-                            })}
-                          />
-                        </Field>
-                      </FieldSet>
-                    </CarouselItem>
-                  )}
-
-                  {/* Organic Question */}
-                  <CarouselItem>
-                    <FieldSet>
-                      <FieldLabel
-                        htmlFor="is-organic"
-                        className="text-lg leading-tight mt-23 md:mt-16 lg:mt-12"
-                      >
-                        {t('questions.organic', { farm: getValues().farmName })}
-                      </FieldLabel>
-
-                      <div className="flex gap-8">
-                        <Button
-                          type="button"
-                          className={`text-lg hover:cursor-pointer decoration-[1px] underline-offset-4 font-thin ${
-                            isOrganic === true
-                              ? 'underline font-normal'
-                              : 'no-underline hover:underline'
-                          }`}
-                          onClick={() => {
-                            setValue('isOrganic', true, {
-                              shouldDirty: true,
-                              shouldValidate: true,
-                            });
-                          }}
-                        >
-                          {t('buttons.yes')}
-                        </Button>
-
-                        <Button
-                          type="button"
-                          className={`text-lg hover:cursor-pointer decoration-[1px] underline-offset-4 font-thin ${
-                            isOrganic === false
-                              ? 'underline font-normal'
-                              : 'no-underline hover:underline'
-                          }`}
-                          onClick={() => {
-                            setValue('isOrganic', false, {
-                              shouldDirty: true,
-                              shouldValidate: true,
-                            });
-                          }}
-                        >
-                          {t('buttons.no')}
-                        </Button>
-                      </div>
-                    </FieldSet>
-                  </CarouselItem>
-
-                  {/* Hydroponic Question */}
-                  <CarouselItem>
-                    <FieldSet>
-                      <FieldLabel
-                        htmlFor="is-hydroponic"
-                        className="text-lg leading-tight mt-23 md:mt-16 lg:mt-12"
-                      >
-                        {t('questions.hydroponic', {
-                          farm: getValues().farmName,
-                        })}
-                      </FieldLabel>
-
-                      <div className="flex gap-8">
-                        <Button
-                          type="button"
-                          className={`text-lg hover:cursor-pointer decoration-[1px] underline-offset-4 font-thin ${
-                            isHydroponic === true
-                              ? 'underline font-normal'
-                              : 'no-underline hover:underline'
-                          }`}
-                          onClick={() => {
-                            setValue('isHydroponic', true, {
-                              shouldDirty: true,
-                              shouldValidate: true,
-                            });
-                          }}
-                        >
-                          {t('buttons.yes')}
-                        </Button>
-
-                        <Button
-                          type="button"
-                          className={`text-lg hover:cursor-pointer decoration-[1px] underline-offset-4 font-thin ${
-                            isHydroponic === false
-                              ? 'underline font-normal'
-                              : 'no-underline hover:underline'
-                          }`}
-                          onClick={() => {
-                            setValue('isHydroponic', false, {
-                              shouldDirty: true,
-                              shouldValidate: true,
-                            });
-                          }}
-                        >
-                          {t('buttons.no')}
-                        </Button>
-                      </div>
-                    </FieldSet>
-                  </CarouselItem>
-
-                  {/* Sprouts Question */}
-                  <CarouselItem>
-                    <FieldSet>
-                      <FieldLabel
-                        htmlFor="produces-sprouts"
-                        className="text-lg leading-tight mt-23 md:mt-10 lg:mt-12"
-                      >
-                        {t('questions.sprouts', { farm: getValues().farmName })}
-                      </FieldLabel>
-
-                      <div className="flex gap-8">
-                        <Button
-                          type="button"
-                          className={`text-lg hover:cursor-pointer decoration-[1px] underline-offset-4 font-thin ${
-                            producesSprouts === true
-                              ? 'underline font-normal'
-                              : 'no-underline hover:underline'
-                          }`}
-                          onClick={() => {
-                            setValue('producesSprouts', true, {
-                              shouldDirty: true,
-                              shouldValidate: true,
-                            });
-                            api?.scrollNext();
-                          }}
-                        >
-                          {t('buttons.yes')}
-                        </Button>
-
-                        <Button
-                          type="button"
-                          className={`text-lg hover:cursor-pointer decoration-[1px] underline-offset-4 font-thin ${
-                            producesSprouts === false
-                              ? 'underline font-normal'
-                              : 'no-underline hover:underline'
-                          }`}
-                          onClick={() => {
-                            setValue('producesSprouts', false, {
-                              shouldDirty: true,
-                              shouldValidate: true,
-                            });
-                            api?.scrollNext();
-                          }}
-                        >
-                          {t('buttons.no')}
-                        </Button>
-                      </div>
-                    </FieldSet>
-                  </CarouselItem>
-
-                  <CarouselItem>
-                    {isMatch ? (
-                      <div className="flex h-full flex-col mt-22 md:mt-4 lg:mt-8 gap-6 items-center md:items-start gap-6 mx-auto">
-                        <h1 className="text-4xl lg:text-5xl">
-                          {t('results.matchTitle')}
-                        </h1>
-                        <p className="text-normal md:text-lg">
-                          {t('results.matchBody')}
-                        </p>
-                        <SubmitButton
-                          className="w-[200px] h-1s1 ml-1 rounded-full mt-4"
-                          buttonText={t('results.matchJoinUs')}
-                          onClickFunction={onSubmit}
-                        />
-                      </div>
-                    ) : (
-                      <div className="flex h-full flex-col mt-22 md:mt-6 lg:mt-10 gap-6 items-center md:items-start mx-auto">
-                        <h1 className="text-2xl md:text-3xl md:max-w-[360px] lg:max-w-[420px]">
-                          {t('results.noMatchTitle')}
-                        </h1>
-                        <p className="text-sm md:text-normal md:max-w-[360px] lg:max-w-[420px]">
-                          {t('results.noMatchBody')}
-                        </p>
-                        {/* WIP: Need form and logic** to submit the Instagram handle. */}
-                        <p className="text-sm md:text-normal max-w-[280px] sm:max-w-[360px] lg:max-w-[420px]">
-                          {t('results.noMatchInstagram')}
-                        </p>
-                        <Input
-                          className="border-[#848484]/80 border-1 w-full max-w-[280px] sm:max-w-[355px]"
-                          type="text"
-                          placeholder="@handle"
-                          {...register('instagramHandle')}
-                        />
-                        <SubmitButton
-                          className="w-[200px] h-11 ml-1 rounded-full mt-2"
-                          buttonText={t('results.noMatchSubmit')}
-                          onClickFunction={onInstagramSubmit}
-                        />
-                      </div>
-                    )}
-                  </CarouselItem>
-                </CarouselContent>
-                <LegalSubtext />
-
-                <div className="flex min-h-10 justify-between flex-col lg:flex-row gap-4 lg:gap-0">
-                  {slide !== 0 && slide !== totalSlides - 1 ? (
-                    <FadeIn>
-                      <Button
-                        variant="outline"
-                        className="rounded-full h-auto px-16 h-11 max-w-45 text-sm hover:cursor-pointer hover:border-[#848484]/80 font-semibold"
-                        type="button"
-                        onClick={handleBack}
-                      >
-                        {t('buttons.back')}
-                      </Button>
-                    </FadeIn>
-                  ) : (
-                    <div />
-                  )}
-
-                  {slide < totalSlides - 2 ? (
-                    <FadeIn>
-                      <Button
-                        variant="brand"
-                        className="rounded-full text-background h-auto px-13 h-11 max-w-45 text-sm hover:cursor-pointer font-semibold"
-                        type="button"
-                        onClick={handleNext}
-                        disabled={!canAdvance}
-                      >
-                        {t('buttons.next')}
-                      </Button>
-                    </FadeIn>
-                  ) : (
-                    <div />
-                  )}
+            <FadeIn>
+              <div className="mx-auto flex flex-col justify-start w-full max-w-[280px] sm:max-w-[450px] md:max-w-[500px]">
+                <div className="flex h-full flex-col md:mt-10 gap-6 items-start lg:max-w-[430px]">
+                  <h1 className="text-2xl lg:text-3xl md:mb-5 text-left">
+                    {isSubmitted ? t('success.title') : t('hero.title')}
+                  </h1>
+                  <p className="text-sm md:text-normal mb-8 text-left font-normal">
+                    {isSubmitted ? t('success.subtitle') : t('hero.subtitle')}
+                  </p>
                 </div>
-              </Carousel>
-            </form>
+                {isSubmitted ? (
+                  <div className="flex justify-start mt-6">
+                    <Button
+                      onClick={() => router.push('/')}
+                      className="w-[144px] rounded-full bg-black text-white hover:cursor-pointer hover:bg-black/80"
+                      type="button"
+                    >
+                      {t('success.cta')}
+                    </Button>
+                  </div>
+                ) : (
+                  <>
+                    {errors.root?.message && (
+                      <FormErrorMessage errorMessage={errors.root.message} />
+                    )}
+                    <form
+                      onSubmit={handleSubmit(onSubmit)}
+                      className="flex flex-col gap-4 mb-24"
+                    >
+                      <FieldSet className="mb-8">
+                        <FieldGroup>
+                          <Field>
+                            <div className="flex flex-row justify-between">
+                              <FieldLabel
+                                htmlFor="name"
+                                className="leading-tight mb-[-6px]"
+                              >
+                                {t('fields.name')}
+                              </FieldLabel>
+                              <ErrorMessage
+                                errors={errors}
+                                name="name"
+                                render={({ message }) => (
+                                  <FormErrorMessage errorMessage={message} />
+                                )}
+                              />
+                            </div>
+                            <Input
+                              className="border-[#848484]/80 border-1"
+                              id="name"
+                              data-testid="name"
+                              {...register('name')}
+                              type="text"
+                              required
+                            />
+                          </Field>
+                          <Field>
+                            <div className="flex flex-row justify-between">
+                              <FieldLabel
+                                htmlFor="lastKnownEmail"
+                                className="leading-tight mb-[-6px]"
+                              >
+                                {t('fields' + emailType)}
+                              </FieldLabel>
+                              <ErrorMessage
+                                errors={errors}
+                                name="lastKnownEmail"
+                                render={({ message }) => (
+                                  <FormErrorMessage errorMessage={message} />
+                                )}
+                              />
+                            </div>
+                            <Input
+                              id="lastKnownEmail"
+                              data-testid="lastKnownEmail"
+                              type="email"
+                              className="border-[#848484]/80 border-1"
+                              {...register('lastKnownEmail')}
+                              required
+                            />
+                          </Field>
+                          <Field>
+                            <div className="flex flex-row justify-between">
+                              <FieldLabel
+                                htmlFor="response"
+                                className="leading-tight mb-[-6px]"
+                              >
+                                {t('fields.message')}
+                              </FieldLabel>
+                              <ErrorMessage
+                                errors={errors}
+                                name="response"
+                                render={({ message }) => (
+                                  <FormErrorMessage errorMessage={message} />
+                                )}
+                              />
+                            </div>
+                            <div className="flex flex-row items-center justify-center gap-2 text-nowrap">
+                              <Textarea
+                                id="response"
+                                className="border-[#848484]/80 border-1 ring-0! focus-visible:ring-0! focus-visible:ring-offset-0!"
+                                {...register('response')}
+                                rows={4}
+                                required
+                              />
+                            </div>
+                          </Field>
+                        </FieldGroup>
+                      </FieldSet>
+                      <SubmitButton
+                        buttonText={t('buttons.submit')}
+                        className="w-[144px] h-11 rounded-full px-8 py-3 font-semibold"
+                        disabled={isSubmitting || !isValid}
+                        reactHookFormPending={isSubmitting}
+                      />
+                    </form>
+                  </>
+                )}
+              </div>
+            </FadeIn>
           </div>
         </div>
       </div>
     </main>
   );
 }
+
+/* ARCHIVED: Inbound onboarding carousel previously served at /contact.
+ * Kept for reference until the account setup flow is finalized.
+ * To restore, swap the active Contact export and update ./types + i18n.
+ */
+// // Copyright ┬⌐ Todd Agriscience, Inc. All rights reserved.
+//
+// 'use client';
+//
+// import { FadeIn } from '@/components/common';
+// import FormErrorMessage from '@/components/common/form-error-message/form-error-message';
+// import { LegalSubtext } from '@/components/common/legal-subtext/legal-subtext';
+// import MarketingGradientBox from '@/components/common/marketing-gradient-box/marketing-gradient-box';
+// import SubmitButton from '@/components/common/utils/submit-button/submit-button';
+// import { Button } from '@/components/ui';
+// import {
+//   Carousel,
+//   CarouselApi,
+//   CarouselContent,
+//   CarouselItem,
+// } from '@/components/ui/carousel';
+// import {
+//   Field,
+//   FieldLabel,
+//   FieldLegend,
+//   FieldSet,
+// } from '@/components/ui/field';
+// import { Input } from '@/components/ui/input';
+// import { submitContactToSheets } from '@/lib/actions/googleSheets';
+// import isWorkEmail from '@/lib/utils/is-work-email';
+// import { ErrorMessage } from '@hookform/error-message';
+// import { zodResolver } from '@hookform/resolvers/zod';
+// import Fade from 'embla-carousel-fade';
+// import { useTranslations } from 'next-intl';
+// import Link from 'next/link';
+// import { useRouter } from 'next/navigation';
+// import { useEffect, useState } from 'react';
+// import { useForm } from 'react-hook-form';
+// import { ContactFormData, contactFormSchema } from './types';
+//
+// function ContactOnboardingCarousel() {
+//   const t = useTranslations('contactPage');
+//   const [slide, setSlide] = useState(0);
+//   const router = useRouter();
+//   const [totalSlides, setTotalSlides] = useState(-1);
+//   const [api, setApi] = useState<CarouselApi>();
+//   const {
+//     register,
+//     trigger,
+//     getValues,
+//     clearErrors,
+//     setValue,
+//     watch,
+//     formState: { errors, isValid },
+//   } = useForm<ContactFormData>({
+//     defaultValues: {
+//       name: undefined,
+//       firstName: '',
+//       lastName: '',
+//       farmName: '',
+//       email: '',
+//       phone: '',
+//       website: undefined,
+//       isOrganic: undefined,
+//       isHydroponic: undefined,
+//       producesSprouts: undefined,
+//       instagramHandle: '',
+//     },
+//     resolver: zodResolver(contactFormSchema),
+//   });
+//
+//   const {
+//     firstName,
+//     lastName,
+//     farmName,
+//     email,
+//     phone,
+//     website,
+//     isOrganic,
+//     isHydroponic,
+//     producesSprouts,
+//     instagramHandle,
+//   } = watch();
+//   const requiresWebsite = !isWorkEmail(email ?? '');
+//   const organicSlideIndex = requiresWebsite ? 2 : 1;
+//   const hydroponicSlideIndex = organicSlideIndex + 1;
+//   const sproutsSlideIndex = organicSlideIndex + 2;
+//   const instagramSlideIndex = organicSlideIndex + 3;
+//   const canAdvance =
+//     (slide === 0 &&
+//       Boolean(firstName && lastName && farmName && email && phone)) ||
+//     (requiresWebsite && slide === 1 && Boolean(website)) ||
+//     (slide === organicSlideIndex && isOrganic !== undefined) ||
+//     (slide === hydroponicSlideIndex && isHydroponic !== undefined) ||
+//     (slide === sproutsSlideIndex && producesSprouts !== undefined) ||
+//     (slide === instagramSlideIndex && Boolean(instagramHandle));
+//
+//   useEffect(() => {
+//     if (!api) return;
+//
+//     setTotalSlides(api.slideNodes().length);
+//     setSlide(api.selectedScrollSnap());
+//
+//     api.on('select', () => setSlide(api.selectedScrollSnap()));
+//     api.on('slidesChanged', () => setTotalSlides(api.slideNodes().length));
+//   }, [api]);
+//
+//   function handleBack() {
+//     api?.scrollPrev();
+//   }
+//
+//   async function handleNext() {
+//     if (!api) return;
+//
+//     const fieldsToValidate =
+//       slide === 0
+//         ? ([
+//             'firstName',
+//             'lastName',
+//             'farmName',
+//             'email',
+//             'phone',
+//           ] as (keyof ContactFormData)[])
+//         : requiresWebsite && slide === 1
+//           ? (['website'] as (keyof ContactFormData)[])
+//           : slide === organicSlideIndex
+//             ? (['isOrganic'] as (keyof ContactFormData)[])
+//             : slide === hydroponicSlideIndex
+//               ? (['isHydroponic'] as (keyof ContactFormData)[])
+//               : slide === sproutsSlideIndex
+//                 ? (['producesSprouts'] as (keyof ContactFormData)[])
+//                 : slide === instagramSlideIndex
+//                   ? (['instagramHandle'] as (keyof ContactFormData)[])
+//                   : [];
+//
+//     const isStepValid = fieldsToValidate.length
+//       ? await trigger(fieldsToValidate)
+//       : true;
+//
+//     if (isStepValid) {
+//       api.scrollNext();
+//       clearErrors();
+//     }
+//   }
+//
+//   // This function is triggered directly by the <SubmitButton/> because onSubmit via the form itself isn't working for some reason.
+//   async function onSubmit() {
+//     await trigger();
+//
+//     if (isValid) {
+//       const params = new URLSearchParams({
+//         first_name: getValues().firstName,
+//         last_name: getValues().lastName,
+//         farm_name: getValues().farmName,
+//         email: getValues().email,
+//         phone: getValues().phone,
+//         instagram_handle: getValues().instagramHandle || '',
+//       });
+//       router.push(`/signup?${params.toString()}`);
+//     }
+//   }
+//
+//   async function onInstagramSubmit() {
+//     await trigger();
+//
+//     if (isValid) {
+//       const params = new URLSearchParams({
+//         first_name: getValues().firstName,
+//         last_name: getValues().lastName,
+//         farm_name: getValues().farmName,
+//         email: getValues().email,
+//         phone: getValues().phone,
+//         instagram_handle: getValues().instagramHandle || '',
+//       });
+//       // Push to Google Sheets
+//       try {
+//         await submitContactToSheets(Object.fromEntries(params.entries()));
+//       } catch (error) {
+//         console.error('Failed to submit to Google Sheets:', error);
+//       }
+//       // Redirect to What We Do
+//       router.push(`/research`);
+//     }
+//   }
+//
+//   const isMatch =
+//     (isWorkEmail(getValues().email) || getValues().website) &&
+//     (getValues().isOrganic || getValues().isOrganic === undefined) &&
+//     !getValues().isHydroponic &&
+//     !getValues().producesSprouts;
+//
+//   return (
+//     <main>
+//       <div className="max-w-[1400px] mx-auto px-15 lg:px-16 flex items-center justify-center">
+//         <div className="grid grid-cols-1 md:grid-cols-2 gap-8 lg:gap-16 place-items-start mx-auto mt-5 md:mt-15 w-full max-w-[1200px] mx-auto">
+//           <MarketingGradientBox />
+//           <div className="flex w-full max-w-[530px] lg:max-w-none flex-col md:mr-0 lg:mr-10">
+//             <form
+//               onSubmit={(e) => e.preventDefault()}
+//               onKeyDown={(e) => {
+//                 if (e.key === 'Enter' && isValid) {
+//                   e.preventDefault();
+//                   api?.scrollNext();
+//                 }
+//               }}
+//             >
+//               <Carousel
+//                 setApi={setApi}
+//                 className="mx-auto flex flex-col justify-start w-full max-w-[300px] sm:max-w-[450px] md:max-w-[500px]"
+//                 plugins={[Fade()]}
+//                 opts={{ duration: 30, watchDrag: false }}
+//               >
+//                 <CarouselContent className="p-0 h-auto">
+//                   {/* Get Started with Todd*/}
+//                   <CarouselItem>
+//                     <FieldSet className="flex flex-col gap-3">
+//                       <FieldLegend>
+//                         <h1 className="text-2xl mb-8 md:mb-2 mt-10 text-left">
+//                           {t('description')}
+//                         </h1>
+//                       </FieldLegend>
+//                       {/** Honeypot */}
+//                       <Field className="hidden">
+//                         <FieldLabel htmlFor="name" className="leading-tight">
+//                           {t('fields.name')}
+//                         </FieldLabel>
+//                         <Input
+//                           className="border-[#848484]/80 border-1 "
+//                           placeholder={t('placeholders.name')}
+//                           type="text"
+//                           {...register('name')}
+//                         />
+//                       </Field>
+//
+//                       <div className="grid grid-cols-1 sm:grid-cols-2 gap-2 sm:gap-4">
+//                         <Field>
+//                           <div className="flex flex-row justify-between">
+//                             <FieldLabel
+//                               htmlFor="first-name"
+//                               className="leading-tight mb-[-6px]"
+//                             >
+//                               {t('fields.firstName')}
+//                             </FieldLabel>
+//                             <ErrorMessage
+//                               errors={errors}
+//                               name="firstName"
+//                               render={({ message }) => (
+//                                 <FormErrorMessage errorMessage={message} />
+//                               )}
+//                             />
+//                           </div>
+//                           <Input
+//                             className="border-[#848484]/80 border-1"
+//                             required
+//                             {...register('firstName', {
+//                               required: 'This field is required.',
+//                             })}
+//                           />
+//                         </Field>
+//
+//                         <Field>
+//                           <div className="flex flex-row justify-between">
+//                             <FieldLabel
+//                               htmlFor="last-name"
+//                               className="leading-tight mb-[-6px]"
+//                             >
+//                               {t('fields.lastName')}
+//                             </FieldLabel>
+//                             <ErrorMessage
+//                               errors={errors}
+//                               name="lastName"
+//                               render={({ message }) => (
+//                                 <FormErrorMessage errorMessage={message} />
+//                               )}
+//                             />
+//                           </div>
+//                           <Input
+//                             className="border-[#848484]/80 border-1"
+//                             required
+//                             {...register('lastName', {
+//                               required: 'This field is required.',
+//                             })}
+//                           />
+//                         </Field>
+//                       </div>
+//
+//                       <Field>
+//                         <div className="flex flex-row justify-between">
+//                           <FieldLabel
+//                             htmlFor="farm-name"
+//                             className="leading-tight mb-[-6px]"
+//                           >
+//                             {t('fields.farmName')}
+//                           </FieldLabel>
+//                           <ErrorMessage
+//                             errors={errors}
+//                             name="farmName"
+//                             render={({ message }) => (
+//                               <FormErrorMessage errorMessage={message} />
+//                             )}
+//                           />
+//                         </div>
+//                         <Input
+//                           className="border-[#848484]/80 border-1"
+//                           required
+//                           {...register('farmName', {
+//                             required: 'This field is required.',
+//                           })}
+//                         />
+//                       </Field>
+//
+//                       <Field>
+//                         <div className="flex flex-row justify-between">
+//                           <FieldLabel
+//                             htmlFor="email"
+//                             className="leading-tight mb-[-6px]"
+//                           >
+//                             {t('fields.email')}
+//                           </FieldLabel>
+//                           <ErrorMessage
+//                             errors={errors}
+//                             name="email"
+//                             render={({ message }) => (
+//                               <FormErrorMessage errorMessage={message} />
+//                             )}
+//                           />
+//                         </div>
+//                         <Input
+//                           className="border-[#848484]/80 border-1"
+//                           type="email"
+//                           required
+//                           {...register('email', {
+//                             required: 'This field is required.',
+//                           })}
+//                         />
+//                       </Field>
+//
+//                       <Field>
+//                         <div className="flex flex-row justify-between">
+//                           <FieldLabel
+//                             htmlFor="phone"
+//                             className="leading-tight mb-[-6px]"
+//                           >
+//                             {t('fields.phone')}
+//                           </FieldLabel>
+//                           <ErrorMessage
+//                             errors={errors}
+//                             name="phone"
+//                             render={({ message }) => (
+//                               <FormErrorMessage errorMessage={message} />
+//                             )}
+//                           />
+//                         </div>
+//                         <Input
+//                           className="border-[#848484]/80 border-1"
+//                           type="tel"
+//                           required
+//                           defaultValue={'+1'}
+//                           {...register('phone', {
+//                             required: 'This field is required.',
+//                           })}
+//                           maxLength={10}
+//                         />
+//                       </Field>
+//                     </FieldSet>
+//                     <div className="mt-10">
+//                       <p className="text-xs mb-5 font-thin">
+//                         Already started?{' '}
+//                         <Link href="/login" className="underline font-normal">
+//                           Login to complete your application
+//                         </Link>
+//                       </p>
+//                     </div>
+//                   </CarouselItem>
+//                   {/* Website Question */}
+//                   {!isWorkEmail(getValues().email) && (
+//                     <CarouselItem>
+//                       <FieldSet>
+//                         <Field>
+//                           <div className="flex flex-row justify-between mt-23 md:mt-16 lg:mt-12">
+//                             <FieldLabel
+//                               htmlFor="website"
+//                               className="text-lg leading-tight"
+//                             >
+//                               {t('questions.website', {
+//                                 farm: getValues().farmName,
+//                               })}
+//                             </FieldLabel>
+//                             <ErrorMessage
+//                               errors={errors}
+//                               name="website"
+//                               render={({ message }) => (
+//                                 <FormErrorMessage errorMessage={message} />
+//                               )}
+//                             />
+//                           </div>
+//                           <Input
+//                             placeholder="https://"
+//                             className="border-[#848484]/80 border-1"
+//                             {...register('website')}
+//                             required
+//                             {...register('website', {
+//                               required: 'This field is required.',
+//                             })}
+//                           />
+//                         </Field>
+//                       </FieldSet>
+//                     </CarouselItem>
+//                   )}
+//
+//                   {/* Organic Question */}
+//                   <CarouselItem>
+//                     <FieldSet>
+//                       <FieldLabel
+//                         htmlFor="is-organic"
+//                         className="text-lg leading-tight mt-23 md:mt-16 lg:mt-12"
+//                       >
+//                         {t('questions.organic', { farm: getValues().farmName })}
+//                       </FieldLabel>
+//
+//                       <div className="flex gap-8">
+//                         <Button
+//                           type="button"
+//                           className={`text-lg hover:cursor-pointer decoration-[1px] underline-offset-4 font-thin ${
+//                             isOrganic === true
+//                               ? 'underline font-normal'
+//                               : 'no-underline hover:underline'
+//                           }`}
+//                           onClick={() => {
+//                             setValue('isOrganic', true, {
+//                               shouldDirty: true,
+//                               shouldValidate: true,
+//                             });
+//                           }}
+//                         >
+//                           {t('buttons.yes')}
+//                         </Button>
+//
+//                         <Button
+//                           type="button"
+//                           className={`text-lg hover:cursor-pointer decoration-[1px] underline-offset-4 font-thin ${
+//                             isOrganic === false
+//                               ? 'underline font-normal'
+//                               : 'no-underline hover:underline'
+//                           }`}
+//                           onClick={() => {
+//                             setValue('isOrganic', false, {
+//                               shouldDirty: true,
+//                               shouldValidate: true,
+//                             });
+//                           }}
+//                         >
+//                           {t('buttons.no')}
+//                         </Button>
+//                       </div>
+//                     </FieldSet>
+//                   </CarouselItem>
+//
+//                   {/* Hydroponic Question */}
+//                   <CarouselItem>
+//                     <FieldSet>
+//                       <FieldLabel
+//                         htmlFor="is-hydroponic"
+//                         className="text-lg leading-tight mt-23 md:mt-16 lg:mt-12"
+//                       >
+//                         {t('questions.hydroponic', {
+//                           farm: getValues().farmName,
+//                         })}
+//                       </FieldLabel>
+//
+//                       <div className="flex gap-8">
+//                         <Button
+//                           type="button"
+//                           className={`text-lg hover:cursor-pointer decoration-[1px] underline-offset-4 font-thin ${
+//                             isHydroponic === true
+//                               ? 'underline font-normal'
+//                               : 'no-underline hover:underline'
+//                           }`}
+//                           onClick={() => {
+//                             setValue('isHydroponic', true, {
+//                               shouldDirty: true,
+//                               shouldValidate: true,
+//                             });
+//                           }}
+//                         >
+//                           {t('buttons.yes')}
+//                         </Button>
+//
+//                         <Button
+//                           type="button"
+//                           className={`text-lg hover:cursor-pointer decoration-[1px] underline-offset-4 font-thin ${
+//                             isHydroponic === false
+//                               ? 'underline font-normal'
+//                               : 'no-underline hover:underline'
+//                           }`}
+//                           onClick={() => {
+//                             setValue('isHydroponic', false, {
+//                               shouldDirty: true,
+//                               shouldValidate: true,
+//                             });
+//                           }}
+//                         >
+//                           {t('buttons.no')}
+//                         </Button>
+//                       </div>
+//                     </FieldSet>
+//                   </CarouselItem>
+//
+//                   {/* Sprouts Question */}
+//                   <CarouselItem>
+//                     <FieldSet>
+//                       <FieldLabel
+//                         htmlFor="produces-sprouts"
+//                         className="text-lg leading-tight mt-23 md:mt-10 lg:mt-12"
+//                       >
+//                         {t('questions.sprouts', { farm: getValues().farmName })}
+//                       </FieldLabel>
+//
+//                       <div className="flex gap-8">
+//                         <Button
+//                           type="button"
+//                           className={`text-lg hover:cursor-pointer decoration-[1px] underline-offset-4 font-thin ${
+//                             producesSprouts === true
+//                               ? 'underline font-normal'
+//                               : 'no-underline hover:underline'
+//                           }`}
+//                           onClick={() => {
+//                             setValue('producesSprouts', true, {
+//                               shouldDirty: true,
+//                               shouldValidate: true,
+//                             });
+//                             api?.scrollNext();
+//                           }}
+//                         >
+//                           {t('buttons.yes')}
+//                         </Button>
+//
+//                         <Button
+//                           type="button"
+//                           className={`text-lg hover:cursor-pointer decoration-[1px] underline-offset-4 font-thin ${
+//                             producesSprouts === false
+//                               ? 'underline font-normal'
+//                               : 'no-underline hover:underline'
+//                           }`}
+//                           onClick={() => {
+//                             setValue('producesSprouts', false, {
+//                               shouldDirty: true,
+//                               shouldValidate: true,
+//                             });
+//                             api?.scrollNext();
+//                           }}
+//                         >
+//                           {t('buttons.no')}
+//                         </Button>
+//                       </div>
+//                     </FieldSet>
+//                   </CarouselItem>
+//
+//                   <CarouselItem>
+//                     {isMatch ? (
+//                       <div className="flex h-full flex-col mt-22 md:mt-4 lg:mt-8 gap-6 items-center md:items-start gap-6 mx-auto">
+//                         <h1 className="text-4xl lg:text-5xl">
+//                           {t('results.matchTitle')}
+//                         </h1>
+//                         <p className="text-normal md:text-lg">
+//                           {t('results.matchBody')}
+//                         </p>
+//                         <SubmitButton
+//                           className="w-[200px] h-1s1 ml-1 rounded-full mt-4"
+//                           buttonText={t('results.matchJoinUs')}
+//                           onClickFunction={onSubmit}
+//                         />
+//                       </div>
+//                     ) : (
+//                       <div className="flex h-full flex-col mt-22 md:mt-6 lg:mt-10 gap-6 items-center md:items-start mx-auto">
+//                         <h1 className="text-2xl md:text-3xl md:max-w-[360px] lg:max-w-[420px]">
+//                           {t('results.noMatchTitle')}
+//                         </h1>
+//                         <p className="text-sm md:text-normal md:max-w-[360px] lg:max-w-[420px]">
+//                           {t('results.noMatchBody')}
+//                         </p>
+//                         {/* WIP: Need form and logic** to submit the Instagram handle. */}
+//                         <p className="text-sm md:text-normal max-w-[280px] sm:max-w-[360px] lg:max-w-[420px]">
+//                           {t('results.noMatchInstagram')}
+//                         </p>
+//                         <Input
+//                           className="border-[#848484]/80 border-1 w-full max-w-[280px] sm:max-w-[355px]"
+//                           type="text"
+//                           placeholder="@handle"
+//                           {...register('instagramHandle')}
+//                         />
+//                         <SubmitButton
+//                           className="w-[200px] h-11 ml-1 rounded-full mt-2"
+//                           buttonText={t('results.noMatchSubmit')}
+//                           onClickFunction={onInstagramSubmit}
+//                         />
+//                       </div>
+//                     )}
+//                   </CarouselItem>
+//                 </CarouselContent>
+//                 <LegalSubtext />
+//
+//                 <div className="flex min-h-10 justify-between flex-col lg:flex-row gap-4 lg:gap-0">
+//                   {slide !== 0 && slide !== totalSlides - 1 ? (
+//                     <FadeIn>
+//                       <Button
+//                         variant="outline"
+//                         className="rounded-full h-auto px-16 h-11 max-w-45 text-sm hover:cursor-pointer hover:border-[#848484]/80 font-semibold"
+//                         type="button"
+//                         onClick={handleBack}
+//                       >
+//                         {t('buttons.back')}
+//                       </Button>
+//                     </FadeIn>
+//                   ) : (
+//                     <div />
+//                   )}
+//
+//                   {slide < totalSlides - 2 ? (
+//                     <FadeIn>
+//                       <Button
+//                         variant="brand"
+//                         className="rounded-full text-background h-auto px-13 h-11 max-w-45 text-sm hover:cursor-pointer font-semibold"
+//                         type="button"
+//                         onClick={handleNext}
+//                         disabled={!canAdvance}
+//                       >
+//                         {t('buttons.next')}
+//                       </Button>
+//                     </FadeIn>
+//                   ) : (
+//                     <div />
+//                   )}
+//                 </div>
+//               </Carousel>
+//             </form>
+//           </div>
+//         </div>
+//       </div>
+//     </main>
+//   );
+// }
+//
