+++ apps/site/src/app/(unauthenticated)/[locale]/(contact)/contact/types.ts
@@ -1,25 +1,24 @@
 // Copyright © Todd Agriscience, Inc. All rights reserved.
 
-import { userInfo } from '@/lib/zod-schemas/onboarding';
-import { z } from 'zod';
+import z from 'zod';
 
-/** Notable fields:
- *
- * @property {string} name - This is a honeypot field.
- * @property {boolean} isOrganic - If the farm isn't purely organic, this should be false.
- * @property {boolean} isHydroponic - If the farm is hydroponic, this should be false (we want farms that are not hydroponic).
- * @property {boolean} producesSprouts - If the farm produces sprouts or other non-medium crops as a main product, this should be false (we want farms that do not produce sprouts). */
-export const inboundOnboardingInfo = z.object({
-  // Yes, this is stupid. See: https://github.com/colinhacks/zod/discussions/2801
-  website: z.string().url().or(z.literal('')),
-  isOrganic: z.boolean().optional(),
-  isHydroponic: z.boolean(),
-  producesSprouts: z.boolean(),
+/** Zod schema for the public inquiry form on `/contact`. */
+export const publicInquirySchema = z.object({
+  name: z
+    .string()
+    .trim()
+    .min(1, 'Name is required.')
+    .max(200, 'Name is too long.'),
+  lastKnownEmail: z
+    .string()
+    .min(1, 'Email is required.')
+    .email('Please enter a valid email.')
+    .max(320, 'Email is too long.'),
+  response: z
+    .string()
+    .min(1, 'Response is required.')
+    .max(1500, 'Response is too long (max 1500 characters).'),
 });
 
-export const contactFormSchema = z.intersection(
-  userInfo,
-  inboundOnboardingInfo
-);
-
-export type ContactFormData = z.infer<typeof contactFormSchema>;
+/** Form values for the public inquiry form on `/contact`. */
+export type PublicInquiryData = z.infer<typeof publicInquirySchema>;
