--- apps/site/src/app/(unauthenticated)/[locale]/(contact)/support/layout.tsx
@@ -1,11 +1,0 @@
-// Copyright © Todd Agriscience, Inc. All rights reserved.
-
-import { Metadata } from 'next';
-
-export const metadata: Metadata = {
-  title: 'Support',
-};
-
-export default function Layout({ children }: { children: React.ReactNode }) {
-  return <>{children}</>;
-}
