+++ apps/site/src/app/(unauthenticated)/[locale]/(marketing)/components/footer/footer.tsx
@@ -30,7 +30,7 @@ const Footer = () => {
     { href: '/research', label: t('links.whatWeDo') },
     { href: '/news', label: t('links.news') },
     { href: '/careers', label: t('links.careers') },
-    { href: '/support', label: t('links.contact') },
+    { href: '/contact', label: t('links.contact') },
     // comment out until we have a foundation page
     // { href: '/foundation', label: t('links.foundation') },
   ];
