+++ apps/site/src/components/common/unauthenticated-header/unauthenticated-header.tsx
@@ -5,7 +5,7 @@ import Link from 'next/link';
 
 /**
  * Minimalistic header component for unauthenticated pages
- * Displays the TODD brand and navigation link to support page
+ * Displays the TODD brand and navigation link to the contact page
  * @returns {JSX.Element} - The unauthenticated header component
  */
 export default function UnauthenticatedHeader() {
@@ -16,7 +16,7 @@ export default function UnauthenticatedHeader() {
           <ToddHeader className="flex min-h-10 flex-row items-center" />
           <nav className="flex items-center">
             <Link
-              href="/en/support"
+              href="/contact"
               className="p-1 text-foreground text-base transition-opacity hover:opacity-70 tracking-tight"
             >
               Help
