+++ apps/site/src/i18n/message-files.ts
@@ -22,5 +22,4 @@ export const messageFiles = [
   'terms',
   'careers',
   'cookiePreferences',
-  'support',
 ];
