--- apps/site/src/messages/support/es.json
@@ -1,27 +1,0 @@
-{
-  "supportPage": {
-    "title": "Ayuda",
-    "hero": {
-      "title": "¿Necesitas Ayuda?",
-      "subtitle": "Estamos aquí para ayudarte a obtener el máximo provecho de nuestra plataforma."
-    },
-    "fields": {
-      "name": "Nombre",
-      "email": "Correo Electrónico",
-      "message": "Mensaje"
-    },
-    "placeholders": {
-      "name": "Nombre",
-      "email": "Correo Electrónico",
-      "message": "¿En qué podemos ayudarte?"
-    },
-    "buttons": {
-      "submit": "Enviar"
-    },
-    "success": {
-      "title": "Mensaje Enviado",
-      "subtitle": "Gracias por contactarnos. Te responderemos pronto.",
-      "cta": "Inicio"
-    }
-  }
-}
